+++ database/migrations/2022_11_20_104134_create_material_task_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->bigIncrements('id');
             $table->integer('task_id');
             $table->integer('material_id');
-            $table->double('quantity')->default();
+            $table->float('quantity')->default();
             $table->timestamp('created_at');
         });
     }
